qwertyui

+++ TabBarTemplate-OC/Main_ViewController.m
@@ -92,6 +92,7 @@
     }
     
     UIImage* cellImage = nil;
+//    SDWebImageManager *sdw = [SDWebImageManager sharedManager];
     SDImageCache *imageCache = [SDImageCache sharedImageCache];
     cellImage = [imageCache imageFromDiskCacheForKey:imageURLString];
     if (cellImage) {
@@ -114,7 +115,7 @@
         cell.label.numberOfLines = 3;
     }
     
-    cell.label.text = [NSString stringWithFormat:@"考得上几分氨基酸掉飞机接收到刷卡机大佛我驾驶的空间就圣诞节快捷 电视剧佛尔圣诞节佛为了撒娇的佛巨额玩傻大姐佛额外撒打发我几个考试费离开圣诞节佛为就 快速机动费文件四大皆空佛为福建啥都看见个is阿见附件-%ld-",indexPath.row];
+    cell.label.text = [NSString stringWithFormat:@"考得上几分氨基酸掉飞机接收到刷卡机大佛我驾驶的空间就圣诞节快捷 电视剧佛尔圣诞节佛为了撒娇的佛巨额玩傻大姐佛额外撒打发我几个考试费离开圣诞节佛为就 快速机动费文件四大皆空佛为福建啥都看见个is阿见附件-%ld-",(long)indexPath.row];
     return  cell;
 }
